requirements.txt
boto3==1.28.12
requests==2.31.0
gunicorn==21.0.0
blinker==1.9.0
click==8.1.8
filelock==3.17.0
Flask==3.1.0
fsspec==2025.2.0
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.2.1
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
sympy==1.13.1
torch>=2.0.0
torchvision>=0.15.0
typing_extensions==4.12.2
Werkzeug==3.1.3
zipp==3.21.0
websockets>=11.0.0
yt-dlp>=2023.12.30

